Be predictable in the setter that is used on beans.
When attempting to find a setter on a bean RSF just uses the first method that matches the desired name, returns void and has a single argument. The JVM doesn’t sort the returned methods from a class so if there are two that match these requirements it’s not deterministic about which one will get found.

The clean(er) solution would have been to throw an error if we found more than one setter, but this would have broken some calling code so instead we sort all the found setters to ensure that the results are at least always the same.

We assume that in most cases we won’t find more than one setter so delay the use of an ArrayList.

Fixes #5